fullstack-app/backend/requirements.txt
fastapi==0.123.4
openai==2.8.1
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
python-multipart==0.0.9
langgraph>=0.2.0
langchain-core>=0.3.0
python-docx
beautifulsoup4
pydub
ffmpeg-python
pytest
httpx
PyJWT
